Week 13
Game is pretty much done
some nice to haves if I have time:

1. would love to have grenades, working with rockets caused a lot of still active bugs, maybe a frag grenade would be best.
2. rockets are super buggy, would probably best to make them explode into several bullets instead of attempt to detect collision

3. Bullets move at the wrong speed when the player camera is moving

Other than that, I'm pretty happy with this week. Upcoming I've got to make a couple maps, and a save file.

(Side note: I was going to make a map parser so I could make maps in bulk, but I don't think that makes sense to do with the time I have left. I would love to make it because if I do make a level editor the map parser will be a HUGE help.)